requests implementation
refactoring

added requests endpoint
